Referring to rc.isFrontEndRequest causes errors when using multiple plugins in Mura FW/1.
I was trying to use my own plugin with slatwall and this setting caused errors, because two plugins are both referencing the rc.  By putting this here we're using the variable.fw local to each plugin instead of the global rc context, since the rc is open for anyone to alter and potentially cause issues with.  

It will be more difficult for first time Slatwall or Mura FW/1 plugin users to successfully or easily write a plugin without having to troubleshoot this.  :) My first commit/fork/patch on github!